Calculate a monthly balance
The application begins by displaying the starting balance for a
checking account.
The application prompts the user to enter the amount for a withdrawal
or deposit.
When the user finishes entering deposits and withdrawals, the
application displays the fees for the month followed by the final
balances for the month.